requirements.txt
flask==2.1.2
librespot==0.0.9
python-dotenv==0.20.0
waitress==2.1.2
